refactor(actions): split pipeline/python.rs into responsibility-scoped siblings
pipeline/python.rs (747 lines) becomes a 67-line thin facade that mounts
four focused siblings via #[path]:

- python_helpers (value_to_py + PyCallableHook)
- python_middleware (PyLoggingMiddleware, PyTimingMiddleware,
  PyAuditMiddleware, PyRateLimitMiddleware) — inner fields upgraded to
  pub(super) so the pipeline module can still construct them
- python_shared (SharedTimingMiddleware / SharedAuditMiddleware /
  SharedRateLimitMiddleware Arc newtypes that implement ActionMiddleware)
- python_pipeline (PyActionPipeline, the Python-facing ToolPipeline)

Every Python class is re-exported so pipeline::python::{PyActionPipeline,
PyLoggingMiddleware, …} keeps working unchanged for downstream callers.
The Shared* newtypes are re-exported under #[cfg(test)] so the existing
python_tests.rs unit tests continue to reference super::python::Shared*
unchanged.

228 + 8 pipeline tests green with default features; workspace cargo check
clean. The python-bindings lib-tests compile errors in events.rs are
pre-existing on HEAD and unrelated to this refactor.
